+++ legacy/src/model/formToState/component-new-edit/response-format-table.jsx
@@ -114,6 +114,7 @@ export const defaultMeasureForm = {
       name: '',
       label: '',
       scope: '',
+      optionFilter: '',
     },
     visHint: RADIO,
     choiceType: CHOICE_TYPE_CODES_LIST,
@@ -506,7 +507,10 @@ const Factory = (initialState = {}, codesListsStore) => {
         }
 
         state[SINGLE_CHOICE] = {
-          [listPath]: codesListsStore[measureState[listPath].id],
+          [listPath]: {
+            ...codesListsStore[measureState[listPath].id],
+            optionFilter: measureState[listPath].optionFilter,
+          },
           visHint: measureState.visHint,
           choiceType: measureState.choiceType,
         };